+++ NEWS.rst
@@ -1,3 +1,8 @@
+Since 2.0.0:
+
+* The push_change_refs option which specified that Zuul refs should be
+  pushed to Gerrit has been removed.
+
 Since 1.3.0:
 
 * The Jenkins launcher is replaced with Gearman launcher.  An internal